Update bug
Fixed bug.
Added overflow: visible for when the size of the application shrinks down in order to be mobile friendly to be extent.

Need to create a mobile navigation tab setup for the Beat and Jam in order to be effective with the design. Perhaps leaving them at the top instead of the bottom since the user will have to scroll down in order to click all of the beats for the beat.